Enable jest-matchers in browser context by splitting utils
