+++ packages/jest-cli/package.json
@@ -17,6 +17,8 @@
     "jest-environment-jsdom": "^18.1.0",
     "jest-haste-map": "^18.1.0",
     "jest-jasmine2": "^18.1.0",
+    "jest-message-util": "^18.1.0",
+    "jest-regex-util": "^18.1.0",
     "jest-resolve": "^18.1.0",
     "jest-resolve-dependencies": "^18.1.0",
     "jest-runtime": "^18.1.0",
