+++ packages/jest-config/package.json
@@ -12,8 +12,8 @@
     "jest-environment-jsdom": "^18.1.0",
     "jest-environment-node": "^18.1.0",
     "jest-jasmine2": "^18.1.0",
+    "jest-regex-util": "^18.1.0",
     "jest-resolve": "^18.1.0",
-    "jest-util": "^18.1.0",
     "jest-validate": "^18.1.0",
     "pretty-format": "^18.1.0"
   }
