+++ packages/jest-config/src/normalize.js
@@ -26,7 +26,7 @@ const crypto = require('crypto');
 const DEFAULT_CONFIG = require('./defaults');
 const path = require('path');
 const Resolver = require('jest-resolve');
-const utils = require('jest-util');
+const utils = require('jest-regex-util');
 const VALID_CONFIG = require('./validConfig');
 const DEPRECATED_CONFIG = require('./deprecated');
 const JSON_EXTENSION = '.json';
