+++ packages/jest-matchers/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "jest-diff": "^18.1.0",
     "jest-matcher-utils": "^18.1.0",
-    "jest-util": "^18.1.0"
+    "jest-message-util": "^18.1.0",
+    "jest-regex-util": "^18.1.0"
   }
 }
