+++ packages/jest-message-util/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "jest-jasmine2",
+  "name": "jest-message-util",
   "version": "18.1.0",
   "repository": {
     "type": "git",
@@ -8,10 +8,7 @@
   "license": "BSD-3-Clause",
   "main": "build/index.js",
   "dependencies": {
-    "graceful-fs": "^4.1.6",
-    "jest-matcher-utils": "^18.1.0",
-    "jest-matchers": "^18.1.0",
-    "jest-message-util": "^18.1.0",
-    "jest-snapshot": "^18.1.0"
+    "chalk": "^1.1.1",
+    "micromatch": "^2.3.11"
   }
 }
