+++ packages/jest-message-util/src/index.js
@@ -1,0 +1,25 @@
+/**
+ * Copyright (c) 2014, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @flow
+ */
+
+'use strict';
+
+const {
+  formatExecError,
+  formatResultsErrors,
+  formatStackTrace,
+} = require('./messages');
+const separateMessageFromStack = require('./separateMessageFromStack');
+
+module.exports = {
+  formatExecError,
+  formatResultsErrors,
+  formatStackTrace,
+  separateMessageFromStack,
+};
