+++ packages/jest-regex-util/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "jest-regex-util",
+  "version": "18.1.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/facebook/jest.git"
+  },
+  "license": "BSD-3-Clause",
+  "main": "build/index.js"
+}
