+++ packages/jest-util/package.json
@@ -13,8 +13,8 @@
     "jest-file-exists": "^17.0.0",
     "jest-mock": "^18.0.0",
     "jest-validate": "^18.1.0",
+    "jest-message-util": "^18.1.0",
     "leven": "^2.0.0",
-    "micromatch": "^2.3.11",
     "mkdirp": "^0.5.1"
   }
 }
